Задача 1 решена ЛР 5 ВА

+++ Вычислительные алгоритмы/Лабораторная 5/Lab5_CA/src/model/IFunctionTaskFirst.java
@@ -1,13 +1,13 @@
 package model;
 
 public interface IFunctionTaskFirst {
-    float f(float _x, float _y);
+    double f(double _x, double _y);
 
     // Для Рунге-Кутта
-    double k1(float _x, float _y, float _h);
-    double k2(float _x, float _y, float _h);
-    double k3(float _x, float _y, float _h);
-    double k4(float _x, float _y, float _h);
+    double k1(double _x, double _y, double _h);
+    double k2(double _x, double _y, double _h);
+    double k3(double _x, double _y, double _h);
+    double k4(double _x, double _y, double _h);
 
     XYContainer methodRungeKutt(float _h);
 }
