+++ Вычислительные алгоритмы/Лабораторная 5/Lab5_CA/src/model/IFunctionTaskSecond.java
@@ -1,8 +1,8 @@
 package model;
 
 public interface IFunctionTaskSecond {
-    double f21(float _x, float _y2);
-    double f22(float _x, float _y1);
+    double f21(double _x, double _y2);
+    double f22(double _x, double _y1);
 
-    XYContainer algoRungeKutt(float _h, float _y10, float _y20);
+    XYContainer algoRungeKutt(double _h, double _y10, double _y20);
 }
