+++ Вычислительные алгоритмы/Лабораторная 5/Lab5_CA/src/model/ModelTaskFirst.java
@@ -3,9 +3,9 @@ package model;
 import java.util.ArrayList;
 
 public class ModelTaskFirst implements IFunctionTaskFirst {
-    private float x0;
-    private float xn;
-    private float y0;
+    private double x0;
+    private double xn;
+    private double y0;
 
     public ModelTaskFirst(float _x0, float _xn, float _y0) {
         this.x0 = _x0;
@@ -14,27 +14,27 @@ public class ModelTaskFirst implements IFunctionTaskFirst {
     }
 
     @Override
-    public float f(float _x, float _y) {
-        return (float) (_y*Math.log(_y)/_x);
+    public double f(double _x, double _y) {
+        return _y*Math.log(_y)/_x;
     }
 
     @Override
-    public double k1(float _x, float _y, float _h) {
+    public double k1(double _x, double _y, double _h) {
         return _h*f(_x, _y);
     }
 
     @Override
-    public double k2(float _x, float _y, float _h) {
+    public double k2(double _x, double _y, double _h) {
         return _h*f(_x + _h/2, (float) (_y + k1(_x, _y, _h) / 2));
     }
 
     @Override
-    public double k3(float _x, float _y, float _h) {
+    public double k3(double _x, double _y, double _h) {
         return _h*f(_x + _h/2, (float) (_y + k2(_x, _y, _h) / 2));
     }
 
     @Override
-    public double k4(float _x, float _y, float _h) {
+    public double k4(double _x, double _y, double _h) {
         return _h*f(_x + _h/2, (float) (_y + k3(_x, _y, _h)));
     }
 
@@ -42,19 +42,19 @@ public class ModelTaskFirst implements IFunctionTaskFirst {
     public XYContainer methodRungeKutt(float _h) {
         float h = _h;
 
-        ArrayList<Float> xs = new ArrayList<>();
-        float x = x0;
+        ArrayList<Double> xs = new ArrayList<>();
+        double x = x0;
         while(x < xn + h/3) {
             xs.add(x);
             x += h;
         }
-        ArrayList<Float> ys = new ArrayList<>();
+        ArrayList<Double> ys = new ArrayList<>();
         ys.add(y0);
 
-        float y;
+        double y;
         for (int i = 0; i < xs.size()-1; i++) {
-            y = (float) (ys.get(i) + (k1(xs.get(i), ys.get(i), h) + 2 * k2(xs.get(i), ys.get(i), h) +
-                    2 * k3(xs.get(i), ys.get(i), h) + k4(xs.get(i), ys.get(i), h) / 6));
+            y = ys.get(i) + (k1(xs.get(i), ys.get(i), h) + 2 * k2(xs.get(i), ys.get(i), h) +
+                    2 * k3(xs.get(i), ys.get(i), h) + k4(xs.get(i), ys.get(i), h) / 6);
 
             ys.add(y);
         }
