+++ Вычислительные алгоритмы/Лабораторная 5/Lab5_CA/src/model/ModelTaskSecond.java
@@ -3,8 +3,8 @@ package model;
 import java.util.ArrayList;
 
 public class ModelTaskSecond implements IFunctionTaskSecond {
-    private float a;
-    private float b;
+    private double a;
+    private double b;
 
     public ModelTaskSecond(float _a, float _b) {
         if(_a <= _b) {
@@ -14,41 +14,41 @@ public class ModelTaskSecond implements IFunctionTaskSecond {
     }
 
     @Override
-    public double f21(float _x, float _y2) {
+    public double f21(double _x, double _y2) {
         return Math.atan(_x*_x + _y2*_y2);
     }
 
     @Override
-    public double f22(float _x, float _y1) {
+    public double f22(double _x, double _y1) {
         return Math.sin(_x + _y1);
     }
 
     @Override
-    public XYContainer algoRungeKutt(float _h, float _y10, float _y20) {
-        ArrayList<Float> xValues = new ArrayList<>();
+    public XYContainer algoRungeKutt(double _h, double _y10, double _y20) {
+        ArrayList<Double> xValues = new ArrayList<>();
         xValues.add(this.a);
-        ArrayList<Float> y1Values = new ArrayList<>();
+        ArrayList<Double> y1Values = new ArrayList<>();
         y1Values.add(_y10);
-        ArrayList<Float> y2Values = new ArrayList<>();
+        ArrayList<Double> y2Values = new ArrayList<>();
         y2Values.add(_y20);
 
-        float x = this.a;
-        float y1 = _y10;
-        float y2 = _y20;
+        double x = this.a;
+        double y1 = _y10;
+        double y2 = _y20;
 
-        float k1, m1, k2, m2, k3, m3, k4, m4;
+        double k1, m1, k2, m2, k3, m3, k4, m4;
         while(x < this.b) {
-            k1 = (float) (_h * f21(x, y2));
-            m1 = (float) (_h * f22(x, y1));
+            k1 = (_h * f21(x, y2));
+            m1 = (_h * f22(x, y1));
 
-            k2 = (float) (_h * f21(x + _h / 2, y2 + m1 / 2));
-            m2 = (float) (_h * f22(x + _h / 2, y1 + k1 / 2));
+            k2 = (_h * f21(x + _h / 2, y2 + m1 / 2));
+            m2 = (_h * f22(x + _h / 2, y1 + k1 / 2));
 
-            k3 = (float) (_h * f21(x + _h / 2, y2 + m2 / 2));
-            m3 = (float) (_h * f22(x + _h / 2, y1 + k2 / 2));
+            k3 = (_h * f21(x + _h / 2, y2 + m2 / 2));
+            m3 = (_h * f22(x + _h / 2, y1 + k2 / 2));
 
-            k4 = (float) (_h * f21(x + _h, y2 + m3));
-            m4 = (float) (_h * f22(x + _h, y1 + k3));
+            k4 = (_h * f21(x + _h, y2 + m3));
+            m4 = (_h * f22(x + _h, y1 + k3));
 
             y1 += (k1 + 2 * k2 + 2 * k3 + k4) / 6;
             y2 += (m1 + 2 * m2 + 2 * m3 + m4) / 6;
