+++ Вычислительные алгоритмы/Лабораторная 5/Lab5_CA/src/model/XYContainer.java
@@ -3,30 +3,30 @@ package model;
 import java.util.ArrayList;
 
 public class XYContainer {
-    private ArrayList<Float> xs;
-    private ArrayList<Float> ys;
-    private ArrayList<Float> ys2;
+    private ArrayList<Double> xs;
+    private ArrayList<Double> ys;
+    private ArrayList<Double> ys2;
 
-    public XYContainer(ArrayList<Float> _x, ArrayList<Float> _y) {
+    public XYContainer(ArrayList<Double> _x, ArrayList<Double> _y) {
         this.xs = _x;
         this.ys = _y;
     }
 
-    public XYContainer(ArrayList<Float> _x, ArrayList<Float> _y, ArrayList<Float> _y2) {
+    public XYContainer(ArrayList<Double> _x, ArrayList<Double> _y, ArrayList<Double> _y2) {
         this.xs = _x;
         this.ys = _y;
         this.ys2 = _y2;
     }
 
-    public ArrayList<Float> getXs() {
+    public ArrayList<Double> getXs() {
         return this.xs;
     }
 
-    public ArrayList<Float> getYs() {
+    public ArrayList<Double> getYs() {
         return this.ys;
     }
 
-    public ArrayList<Float> getYs2() {
+    public ArrayList<Double> getYs2() {
         return this.ys2;
     }
 }
